Fix: Remove helm-docs generated changelog
This was causing a huge issue because I refuse to require a forked
version of helm-docs. Honestly, I don't even care about a changelog that
is exposed via markdown for my personal repo. I'll know what changed and
if I need a reminder, that's what Git and commit messages are for.